backend/requirements.txt
Django==4.2.24
djangorestframework==3.16.1
psycopg2-binary==2.9.10
redis==6.4.0
celery==5.5.3
langchain==0.3.27
openai==1.105.0
sentence-transformers==5.1.0
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework-simplejwt==5.3.0
django-redis==5.4.0
pgvector==0.2.4
numpy==2.0.2
scikit-learn==1.6.1
torch==2.8.0
transformers==4.56.0
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.2
python-pptx==0.6.23
aiohttp==3.9.1
httpx==0.25.2
reportlab==4.0.4
openpyxl==3.1.2
# BIM/IFC 转 JSON（上传 .ifc 文件时需要）
ifcopenshell>=0.8.0
